rtl/dxa + simx/tcu_ld: fix slot/wid confusion and 64-bit TCU_LD address
Two related correctness fixes that together unblock NT=8+/XLEN=64
sparse and multi-CTA-resident regression configs:

1. VX_mem_unit DXA receiver translator: the upper NW_WIDTH bits of
   bar_addr encode the receiver's **CTA slot** (= local_group_id,
   what `bar.id()` packs from the kernel via `(user_id<<8) |
   get_local_group_id()`), NOT a physical warp ID. Multicast replay
   also iterates over CTA slots via `replay_next_idx`. The prior code
   read `wid_to_lmem_base[recv_slot]` and silently aliased for
   single-CTA-resident configs (slot 0 == wid 0); with NT=8 / NW=4 /
   2 warps-per-CTA the two co-resident CTAs land at slot 0 and 1,
   and `wid_to_lmem_base[1]` returns CTA-0's wid-1 base — so CTA-1's
   DXA writes corrupt CTA-0's SMEM region. Switch to
   `slot_to_lmem_base[recv_slot]` (same single registered indexed
   MUX, no cascaded chain). Verified on sgemm2_dxa NT=8 with
   N=8/16/64 and on sgemm2_dxa_mcast (no regression).

2. simx tcu_unit::TCU_LD: rs1 carries a full-width address. The
   reader cast it through `.u32`, silently dropping the upper 32
   bits on XLEN=64 and reading the wrong meta buffer when the
   meta allocation lands above 4 GiB. Read as `.u64`.

+++ sim/simx/tcu/tcu_unit.cpp
@@ -542,7 +542,11 @@ public:
       #ifdef VX_CFG_TCU_SPARSE_ENABLE
         case TcuType::TCU_LD: {
           // Warp-broadcast: all lanes hold the same rs1; take lane 0.
-          uint32_t base_addr = rs1_data.empty() ? 0 : rs1_data.at(0).u32;
+          // rs1 is a full-width address (.u64) — truncating to .u32 silently
+          // drops the upper 32 bits on XLEN=64 and reads the wrong meta
+          // buffer (regression: sgemm_tcu_wg_sp x64 produced all-wrong
+          // outputs when the meta buffer landed above 4 GiB).
+          uint64_t base_addr = rs1_data.empty() ? 0 : rs1_data.at(0).u64;
           this->tcu_ld(wid, tpuArgs.fmt_s, tpuArgs.fmt_d, base_addr);
         } break;
       #endif
